Move date and number parsing into utility library.
